+++ lib/cylc/cfgspec/suite.py
@@ -19,13 +19,11 @@
 
 from isodatetime.data import Calendar
 
-from cylc.parsec.upgrade import upgrader
-from cylc.parsec.config import ParsecConfig
-
-from cylc.cfgvalidate import (
-    cylc_config_validate, CylcConfigValidator as VDR, DurationFloat)
 from cylc.network import Priv
-
+from cylc.parsec.config import ParsecConfig
+from cylc.parsec.upgrade import upgrader
+from cylc.parsec.validate import (
+    DurationFloat, CylcConfigValidator as VDR, cylc_config_validate)
 
 # Nested dict of spec items.
 # Spec value is [value_type, default, allowed_2, allowed_3, ...]
